+++ History.md
@@ -1,3 +1,8 @@
+Unpublished
+===========
+
+  * interrupt support with gpioSetISRFunc
+
 0.0.0 / Oct 14 2015
 ===================